Add all rest routes on base path (#279)
* Add new route using blueprints
* Make rabbitmq tests work
* Give _stuf function its own endpoint name
* Make tests initialize rabbitmq
