+++ src/.jenkins/test/docker-compose.yml
@@ -7,5 +7,14 @@ services:
       target: test
     environment:
       GOB_SHARED_DIR: /app/shared
+      MESSAGE_BROKER_ADDRESS: rabbitmq
     command: >
       bash -c "./test.sh"
+    depends_on:
+      - rabbitmq
+  rabbitmq:
+    image: "rabbitmq:3.9-management"
+    environment:
+      RABBITMQ_CONFIG_FILE: /etc/rabbitmq/rabbitmq.conf
+    volumes:
+      - "./rabbitmq.conf:/etc/rabbitmq/rabbitmq.conf"
