+++ src/gobstuf/blueprints/__init__.py
@@ -1,0 +1,2 @@
+from gobstuf.blueprints.hc import hc_bp  # noqa: F401
+from gobstuf.blueprints.secure import secure_bp  # noqa: F401
