+++ src/gobstuf/config.py
@@ -29,7 +29,10 @@ ROUTE_SCHEME = _getenv("ROUTE_SCHEME", default_value="https")
 PKCS12_FILENAME = _getenv("PKCS12_FILENAME", is_optional=True)
 PKCS12_PASSWORD = _getenv("PKCS12_PASSWORD", is_optional=True)
 
-API_BASE_PATH = _getenv("BASE_PATH", default_value="", is_optional=True)
+API_BASE_PATH = _getenv("API_BASE_PATH", default_value="", is_optional=True)
+HC_BASE_PATH = _getenv("HC_BASE_PATH", default_value="", is_optional=True)
+
+BASE_PATHS = [API_BASE_PATH, HC_BASE_PATH]
 
 AUDIT_LOG_CONFIG = {
     'EXEMPT_URLS': [],
