+++ src/gobstuf/rest/routes.py
@@ -10,18 +10,44 @@ from gobstuf.rest.brp.views import (
 )
 
 REST_ROUTES = [
-    ('/brp/ingeschrevenpersonen', IngeschrevenpersonenFilterView.as_view('brp_ingeschrevenpersonen_list')),
-    ('/brp/ingeschrevenpersonen/<bsn>', IngeschrevenpersonenBsnView.as_view('brp_ingeschrevenpersonen_bsn')),
-    ('/brp/ingeschrevenpersonen/<bsn>/partners',
-     IngeschrevenpersonenBsnPartnerListView.as_view('brp_ingeschrevenpersonen_bsn_partners_list')),
-    ('/brp/ingeschrevenpersonen/<bsn>/partners/<partners_id>',
-     IngeschrevenpersonenBsnPartnerDetailView.as_view('brp_ingeschrevenpersonen_bsn_partners_detail')),
-    ('/brp/ingeschrevenpersonen/<bsn>/ouders',
-     IngeschrevenpersonenBsnOudersListView.as_view('brp_ingeschrevenpersonen_bsn_ouders_list')),
-    ('/brp/ingeschrevenpersonen/<bsn>/ouders/<ouders_id>',
-     IngeschrevenpersonenBsnOudersDetailView.as_view('brp_ingeschrevenpersonen_bsn_ouders_detail')),
-    ('/brp/ingeschrevenpersonen/<bsn>/kinderen',
-     IngeschrevenpersonenBsnKinderenListView.as_view('brp_ingeschrevenpersonen_bsn_kinderen_list')),
-    ('/brp/ingeschrevenpersonen/<bsn>/kinderen/<kinderen_id>',
-     IngeschrevenpersonenBsnKinderenDetailView.as_view('brp_ingeschrevenpersonen_bsn_kinderen_detail')),
+    (
+        '/brp/ingeschrevenpersonen',
+        IngeschrevenpersonenFilterView.as_view('brp_ingeschrevenpersonen_list'),
+        ["GET"]
+    ),
+    (
+        '/brp/ingeschrevenpersonen/<bsn>',
+        IngeschrevenpersonenBsnView.as_view('brp_ingeschrevenpersonen_bsn'),
+        ["GET"]
+    ),
+    (
+        '/brp/ingeschrevenpersonen/<bsn>/partners',
+        IngeschrevenpersonenBsnPartnerListView.as_view('brp_ingeschrevenpersonen_bsn_partners_list'),
+        ["GET"]
+    ),
+    (
+        '/brp/ingeschrevenpersonen/<bsn>/partners/<partners_id>',
+        IngeschrevenpersonenBsnPartnerDetailView.as_view('brp_ingeschrevenpersonen_bsn_partners_detail'),
+        ["GET"]
+    ),
+    (
+        '/brp/ingeschrevenpersonen/<bsn>/ouders',
+        IngeschrevenpersonenBsnOudersListView.as_view('brp_ingeschrevenpersonen_bsn_ouders_list'),
+        ["GET"]
+    ),
+    (
+        '/brp/ingeschrevenpersonen/<bsn>/ouders/<ouders_id>',
+        IngeschrevenpersonenBsnOudersDetailView.as_view('brp_ingeschrevenpersonen_bsn_ouders_detail'),
+        ["GET"]
+    ),
+    (
+        '/brp/ingeschrevenpersonen/<bsn>/kinderen',
+        IngeschrevenpersonenBsnKinderenListView.as_view('brp_ingeschrevenpersonen_bsn_kinderen_list'),
+        ["GET"]
+    ),
+    (
+        '/brp/ingeschrevenpersonen/<bsn>/kinderen/<kinderen_id>',
+        IngeschrevenpersonenBsnKinderenDetailView.as_view('brp_ingeschrevenpersonen_bsn_kinderen_detail'),
+        ["GET"]
+    ),
 ]
