+++ src/tests/conftest.py
@@ -6,6 +6,7 @@ import jwt
 import pytest
 from flask import Flask
 from flask.testing import FlaskClient
+from gobcore.message_broker.initialise_queues import initialize_message_broker
 
 from gobcore.secure.request import ACCESS_TOKEN_HEADER, USER_NAME_HEADER
 from gobstuf.api import get_flask_app
@@ -15,7 +16,7 @@ def _env(var: str) -> str:
     return os.environ[var]
 
 
-@pytest.fixture
+@pytest.fixture(scope='module')
 def app() -> Generator[Flask, None, None]:
     """Creates a flask test app, with an app context.
 
@@ -23,6 +24,7 @@ def app() -> Generator[Flask, None, None]:
     with app.test_client() as client:
         client.get(...)
     """
+    initialize_message_broker()
     app = get_flask_app()
     with app.app_context():
         yield app
@@ -117,8 +119,8 @@ def stuf_310_response(request, requests_mock, tests_dir: Path) -> None:
     requests_mock.post(url, text=mock_response)
 
 
-@pytest.fixture
-def app_base_path() -> str:
+@pytest.fixture(params=["API_BASE_PATH", "HC_BASE_PATH"])
+def app_base_path(request) -> str:
     """Returns the base path where the api is mounted.
 
     For example:
@@ -126,4 +128,4 @@ def app_base_path() -> str:
 
     :returns: the path as a string.
     """
-    return os.environ['BASE_PATH']
+    return os.environ[request.param]
