+++ src/tests/pytest.ini
@@ -1,7 +1,8 @@
 [pytest]
 addopts = --capture=no
 env =
-    BASE_PATH=/gob_stuf_test
+    API_BASE_PATH=/gob_stuf_test
+    HC_BASE_PATH=
     ROUTE_PATH_310=/310
     ROUTE_PATH_204=/204
     ROUTE_SCHEME=http
